Deleted browsercomparison.csv
Ended up deleting the browsercomparison.csv (at least temporarily) so that quicker updates to the .json can be pushed (previously had to update by hand - which is why the versions were mismatched. 

For users interested in a temporary work around the table as it lives on https://avoidthehack.com/util/browser-comparison can be copied row by row using the "Copy" button at the top of the table.

Open to a better workflow for the csv file in the future.